to apply service pattern rename legacy module
