+++ studio_app/helpers/__init__.py
@@ -1,5 +1,5 @@
-from ..helpers import *
-from ..helpers import validate_recaptcha
+from ..helpers_legacy import *
+from ..helpers_legacy import validate_recaptcha
 from .email import Email_helper
 __all__ = [
     'email_service',
